# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewsFlow API",
        "description": "Documentação oficial da API NewsFlow",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza login do usuário e retorna token de autenticação",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "responses": {
                    "200": {
                        "description": "Login realizado com sucesso"
                    },
                    "401": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Autenticação"
                ],
                "summary": "Realiza logout do usuário autenticado",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout realizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista posts com filtros opcionais por categoria, tag e título",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "Posts listados com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cria um novo post",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibe um post específico com autor e categoria",
                "operationId": "142b5adb0520cca80cedfe09d9298332",
                "responses": {
                    "200": {
                        "description": "Post encontrado com sucesso"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualiza um post existente",
                "operationId": "650f4bed0846d9350903d8a17c7b503a",
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para atualizar este post"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Remove um post",
                "operationId": "9d99f2e42af2f722a058a86245aa856f",
                "responses": {
                    "200": {
                        "description": "Post deletado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para deletar este post"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
        "/login": {
            "get": {
                "tags": [
                    "Sessão Web"
                ],
                "summary": "Exibe a página de login",
                "operationId": "d38b0625602ec8e41c3bb98aa2e969de",
                "responses": {
                    "200": {
                        "description": "Página de login carregada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sessão Web"
                ],
                "summary": "Realiza login via sessão web",
                "operationId": "42f77e072dec84b0e1094cca1228298e",
                "responses": {
                    "302": {
                        "description": "Login realizado e redirecionamento para dashboard"
                    },
                    "422": {
                        "description": "Credenciais inválidas"
                    }
                }
            }
        },
...
        "/confirm-password": {
            "get": {
                "tags": [
                    "Confirmação de Senha"
                ],
                "summary": "Exibe a página de confirmação de senha",
                "operationId": "cb7fda87fdc05e9ad0ee3b000f4febaa",
                "responses": {
                    "200": {
                        "description": "Página de confirmação carregada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Confirmação de Senha"
                ],
                "summary": "Confirma a senha do usuário autenticado",
                "operationId": "dc62efc6d3a9c874a5c5c974e0b69df4",
                "responses": {
                    "302": {
                        "description": "Senha confirmada com sucesso e redirecionamento"
                    },
                    "422": {
                        "description": "Senha inválida"
                    }
                }
            }
        },
        "/email/verification-notification": {
            "post": {
                "tags": [
                    "Verificação de Email"
                ],
                "summary": "Envia ou reenvia o link de verificação de email para o usuário autenticado",
                "operationId": "7c01a354a5960445207bc5d2f17d547e",
                "responses": {
                    "302": {
                        "description": "Link de verificação enviado ou redirecionamento caso email já esteja verificado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/verify-email": {
            "get": {
                "tags": [
                    "Verificação de Email"
                ],
                "summary": "Exibe a página de verificação de email ou redireciona se já estiver verificado",
                "operationId": "2b8625c52fabf6e7e433cf5a558e1cf2",
                "responses": {
                    "200": {
                        "description": "Página de verificação exibida com sucesso"
                    },
                    "302": {
                        "description": "Usuário já verificado e redirecionado"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
        "/dashboard": {
            "get": {
                "tags": [
                    "Dashboard"
                ],
                "summary": "Exibe o dashboard com posts recentes e categorias",
                "operationId": "a92dff3ea055c0a594f425e852840a68",
                "responses": {
                    "200": {
                        "description": "Dashboard carregado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/posts": {
            "get": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Lista posts paginados no painel web",
                "operationId": "cde581b3119bfed4f4a0ac619622cf97",
                "responses": {
                    "200": {
                        "description": "Posts listados com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Cria um novo post via painel web",
                "operationId": "1affef85a1c5c202305fcb94bafa157e",
                "responses": {
                    "302": {
                        "description": "Post criado com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
...
        "/posts/{id}": {
            "put": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Atualiza um post existente via painel web",
                "operationId": "e11abff47cfc4ec2029a65279ab9808f",
                "responses": {
                    "302": {
                        "description": "Post atualizado com sucesso"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Remove um post via painel web",
                "operationId": "9f9e260077be9abbec62130c26817dea",
                "responses": {
                    "302": {
                        "description": "Post deletado com sucesso"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
...
    },
...
}